requirements.txt
flask
flask-graphql
graphqlclient
graphene
pytest
snapshottest
flake8